requirements.old.txt
Django==4.1.1
django-storages==1.11.1
djangorestframework==3.13.1
gunicorn==20.1.0
azure-common==1.1.28
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
